Add a parameter to fetch to include properties not yet available
Some registers take some time to be available and reading them throws an
ACK modbus exception. This happens for example after powering on the
bridge until it heard RF frames from devices.

But HomeAssistant needs all properties to populate the entities, so add
a parameter to add all device properties to the data dictionary even if
its value is not yet available. The result will be None in these cases.

Signed-off-by: Samuel Cabrero <scabrero@orica.es>